fix: rescan

+++ src/rescan.ts
@@ -1,6 +1,7 @@
 import { dirname } from "node:path";
 import type { MatcherContext } from "./patterns/matcher.js";
 import { scan, type ScanOptions } from "./scan.js";
+import { getDepth } from "./getdepth.js";
 
 /**
  * Returns `true` if the path is included.
@@ -45,48 +46,66 @@ export async function matcherContextAddPath(
  * Removes a path from the {@link MatcherContext.paths} and updates statistics.
  * If you want to remove a source, use {@link matcherContextRefreshDir} instead.
  */
-export async function matcherContextRemovePath(ctx: MatcherContext, entry: string): Promise<void> {
+export async function matcherContextRemovePath(
+  ctx: MatcherContext,
+  entry: string,
+  options: Pick<ScanOptions, "depth">,
+): Promise<void> {
+  if (entry === ".") {
+    entry = "./";
+  }
   const isDir = entry.endsWith("/");
 
   if (!isDir) {
-    ctx.external.delete(dirname(entry));
     if (ctx.paths.delete(entry)) {
+      const parent = dirname(entry);
+      if (ctx.external.get(parent)?.path === entry) {
+        ctx.external.delete(dirname(entry));
+      }
       ctx.totalFiles--;
       ctx.totalMatchedFiles--;
     }
     return;
   }
 
+  entry = entry.substring(0, entry.length - 1);
+
   // if directory
-  ctx.external.delete(entry);
-  if (ctx.paths.delete(entry)) {
+  ctx.external.delete(entry + "/");
+  if (ctx.paths.delete(entry + "/")) {
     ctx.totalDirs--;
   }
 
-  const parent = dirname(entry);
-  const count = (ctx.depthPaths.get(parent) ?? 1) - 1;
-  if (count > 0) {
-    ctx.depthPaths.set(parent, count);
+  const { depth: maxDepth = Infinity } = options;
+  const { depthSlash, depth } = getDepth(entry, maxDepth);
+  if (depth > 0) {
+    const parent = entry.substring(0, depthSlash);
+    const count = (ctx.depthPaths.get(parent) ?? 1) - 1;
+    if (count > 0) {
+      ctx.depthPaths.set(parent, count);
+    }
+    ctx.depthPaths.delete(entry);
   }
-  ctx.depthPaths.delete(entry);
 
   const a = [...ctx.paths];
   for (const path of a) {
-    if (!path.startsWith(entry)) {
+    if (depth > 0 && !path.startsWith(entry)) {
       continue;
     }
 
     const isDir = path.endsWith("/");
     if (!isDir) {
-      if (!path.substring(entry.length).includes("/")) {
-        // for own files
-        ctx.totalFiles--;
-        ctx.totalMatchedFiles--;
+      let subpath = path;
+      if (depth > 0) {
+        subpath = path.substring(entry.length);
+      }
+      if (!subpath.includes("/")) {
+        await matcherContextRemovePath(ctx, path, options);
       }
       continue;
     }
 
-    await matcherContextRemovePath(ctx, path);
+    await matcherContextRemovePath(ctx, path, options);
   }
 }
 
@@ -103,7 +122,7 @@ export async function matcherContextRefreshDir(
     dir = dir.substring(0, dir.length - 1);
   }
 
-  await matcherContextRemovePath(ctx, dir + "/");
+  await matcherContextRemovePath(ctx, dir + "/", options);
 
   const { cwd: cwdo = (await import("node:process")).cwd().replaceAll("\\", "/") } = options;
   let cwd = cwdo + "/" + dir;
